Added some modularity to PNML parsing to add the posibility to add new PNML parsers for some other dialects different from TINA's
